Extract resolver pass
